Fix v_dot4 op_sel modifiers (#2110)
## Motivation

op_sel modifiers are not supported on v_dot4 and v_dot8 instructions.

## Technical Details

Tensile codegen using op_sel for v_dot4 is blocking an llvm patch to
remove the modifiers. They are not supported, and not covered in the ISA
documentation.

## Test Plan

Run regular CI test suite, which includes 4xi8 kernels with the affected
instructions.

## Test Result

Related tests passed when ran locally.

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.