[rocsparse] Fix SpMV sliced ELL rounding errors. (#2940)
## Motivation

Redefine sparse matrix values to be between -1 and 1 for spmv sliced ell
when not converting to int and when converting to int allow using a
range that can be specified in the yaml files using `rand_gen_min` and
`rand_gen_max`

## Technical Details

Seeing rounding errors because input sparse matrix data covers a very
large floating point range of numbers and differences between the host
and device solution math, arch differences, and atomic add usage can
result in failures from inadequate tolerance when comparing the host and
device solution. Solution is to use sparse matrix values constrained
either to a tighter window around 0 or to convert to int as integers can
be exactly represented in float up to 2^24.

## Test Plan

Ensure all tests pass