[HipDNN]: Update dockerfile to be able to use a command line specified rocm libraries commit. (#2947)
## Motivation

The rock isnt always up to date with the head of rocm-libraries. We need
to be able to build dev containers that have the latest regardless of
whats in the rock. Old way to do this was create a branch in TheRock,
now we can just specify the sha during the full build.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.