[hipDNN] [ALMIOPEN-509] Remove workaround and reduce test sizes (#3006)
## Motivation

We were seeing Windows randomly timing out due to large runtime variance
(5 - 10+ minute runs).
A workaround to disable the slowest running tests was put in last week.
After reviewing the tests, this set was also taking a majority of the
runtime for Linux as well.

This change reduces the large test sizes to improve overall runtime, and
ensure they aren't taking > 1/2 of all test time for MIOpen plugin
tests.