[hipDNN] [ALMIOPEN-329] hipDNN using TheRock ASAN build parameters (#3008)
## Motivation
Adjusting our sanitizer cmake builds to work both standalone and in the
context of TheRock. This should allow TheRock ASAN CI builds to also
build hipDNN as ASAN, without losing the ability to do quick ASAN checks
in standalone on our own.

## Technical Details
- Feed off THEROCK_SANITIZER build flag
- Normally, I would do flag forwarding via TheRock's
`ml-libs/CMakeLists.txt`, but it was getting needlessly complex, so
we're going to let some TheRock stuff in for now.

## Test Plan
- Build ASAN using TheRock
- Build ASAN using standalone
- Run all tests

## Test Result
- [x] All tests passing successfully

## Submission Checklist
- [x] pre-commit checks pass