[rocsparse][hipsparse] Update CHANGELOG and version numbers so that they are up-to-date. (#3010)
## Motivation

Update CHANGELOG and version numbers so that they are up-to-date with
currently merged develop branch work

## Technical Details

Some work like the compute residual and addition of sliced ELL format to
hipsparseSpMV was previously intended for 7.2 release but did not make
it in. We therefore need to update develop branches CHANGELOG and
version number.

---------

Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>
Co-authored-by: Kim Liegeois <kimliegeois@ymail.com>