[hipDNN] [ALMIOPEN-322] Don't expand function parameters if log will be dropped. (#3027)
Reduce overhead of logging functions when the log level is disabled.

Also added a small developer enhancement in .gitignore to exclude any
`projects/hipdnn/CMakeUserPresets.json` files that are present in the
workspace.

Previously, when calling the log function, all parameters would be
expanded and passed to the function, only to return without logging if
the log level was disabled.

With the change in this PR, the log level will be checked first, and
then the parameters expanded to the log call only when the log level is
enabled.