[rocRoller] Add colourByNaryArgument utility. (#3490)
## Motivation

To facilitate ordering for the loads of Multiply arguments, we need to
identity if the loads will be used for the LHS, RHS, LHS_SCALE, and/or
RHS_SCALE of the Multiplies.

The `colourByNaryArgument` helper does this.

## Technical Details

To write `colourByNaryArgument`, I also added
`getCoordinateDependencies` to the `ControlFlowRWTracer`. This cycles
through the trace and builds, for each coordinate, a set of coordinates
that it depends upon.

## Test Plan

To test `colourByNaryArgument` I need to add scaling support to the GEMM
common graph.

I also added unit tests for the `ControlFlowRWTracer`. This turned out
to be the bulk of the PR.

## Test Result

This helper is currently not used outside of unit tests. No performance
impact is expected.