[Ex CI] Disabling the MIOpen Azure Pipeline (#3424)
## Motivation
As Azure pipelines are being deprecated, disabling pipelines that are
not needed anymore.
For MIOpen: For this PR [[miopen] Build CK from monorepo for Docker
builds](https://github.com/ROCm/rocm-libraries/pull/3026), we need to
either disable the MIOpen pipeline or change the pipeline to use CK as a
dependency like other projects by being added to the deps array. So we
will instead disable the pipeline.
<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.